+++ R/optim_apsimx.R
@@ -16,6 +16,11 @@
 #' to the initial values of the parameters. So, for example, if your initial value is 20 
 #' and you provide an upper bound of 5, it means that the actual upper value that you are allowing for is 100. 
 #' 
+#' * I have tested other optimizers and packages, but I think these are enough for most purposes. I tried
+#' function stats::nlm (but it does not support bounds and it can fail), package 'optimx' is a bit messy and it
+#' does not provide sufficient additional functionality. Package 'ucminf' seems like a good alternative, but it
+#' did not perform better than the other ones.
+#' 
 #' @title Optimize parameters in an APSIM Next Generation simulation
 #' @name optim_apsimx
 #' @rdname optim_apsimx
@@ -25,7 +30,8 @@
 #' @param parm.paths absolute or relative paths of the coefficients to be optimized. 
 #'             It is recommended that you use \code{\link{inspect_apsimx}} for this
 #' @param data data frame with the observed data. By default is assumes there is a 'Date' column for the index.
-#' @param type Type of optimization. For now, \code{\link[stats]{optim}} and, if available, \code{\link[nloptr]{nloptr}} or \sQuote{mcmc} through \code{\link[BayesianTools]{runMCMC}}.
+#' @param type Type of optimization. For now, \code{\link[stats]{optim}}, and, if available, \code{\link[nloptr]{nloptr}} or 
+#' \sQuote{mcmc} through \code{\link[BayesianTools]{runMCMC}}. Option \sQuote{ucminf} uses the \code{\link[ucminf]{ucminf}} function.
 #' @param weights Weighting method or values for computing the residual sum of squares. 
 #' @param index Index for filtering APSIM output. Typically, \dQuote{Date}, but it can be c(\dQuote{report}, \dQuote{Date}) for 
 #' multiple simulations
@@ -53,7 +59,7 @@
 
 optim_apsimx <- function(file, src.dir = ".", 
                          parm.paths, data, 
-                         type = c("optim", "nloptr", "mcmc"), 
+                         type = c("optim", "nloptr", "mcmc", "ucminf"), 
                          weights, 
                          index = "Date",
                          parm.vector.index,
@@ -91,8 +97,15 @@ optim_apsimx <- function(file, src.dir = ".",
     }
   }
   
+  if(type == "ucminf"){
+    if(!requireNamespace("ucminf", quietly = TRUE)){
+      warning("The ucminf package is required for this method")
+      return(NULL)
+    }
+  }
+  
   ## Setting up Date
-  datami <- data[,-which(names(data) %in% index), drop = FALSE]
+  datami <- data[ ,-which(names(data) %in% index), drop = FALSE]
   if(any(grepl("Date", index))) data$Date <- as.Date(data$Date)
   
   ## Setting up weights
@@ -100,12 +113,12 @@ optim_apsimx <- function(file, src.dir = ".",
     weights <- rep(1, ncol(datami))
   }else{
     if(weights == "mean"){
-      vmns <- abs(1 / apply(datami, 2, mean))  
-      weights <- (vmns / sum(vmns)) * ncol(datami)
+      vmns <- abs(1 / apply(datami, 2, mean, na.rm = TRUE))  
+      weights <- (vmns / sum(vmns)) 
     }else{
       if(weights == "var"){
-        vvrs <- 1 / apply(datami, 2, var)    
-        weights <- (vvrs / sum(vvrs)) * ncol(datami)
+        vvrs <- 1 / apply(datami, 2, var, na.rm = TRUE)    
+        weights <- (vvrs / sum(vvrs)) 
       }else{
         if(length(weights) != ncol(datami))
           stop("Weights not of correct length")
@@ -249,20 +262,20 @@ optim_apsimx <- function(file, src.dir = ".",
     ## Now I need to calculate the residual sum of squares
     ## For this to work all variables should be numeric
     diffs <- as.matrix(data) - as.matrix(sim.s)
-    rss <- sum((diffs * weights)^2)
-    return(rss)
+    rss <- sum(weights * colSums(diffs^2, na.rm = TRUE)) 
+    return(log(rss))
   }
   
   ## Pre-optimized RSS
-  rss <- obj_fun(cfs = rep(1, length(parm.paths)),
-                 parm.paths = parm.paths,
-                 data = data,
-                 iparms = iparms,
-                 weights = weights,
-                 index = index,
-                 parm.vector.index = parm.vector.index,
-                 replacement = replacement,
-                 root = root)
+  pre.lrss <- obj_fun(cfs = rep(1, length(parm.paths)),
+                      parm.paths = parm.paths,
+                      data = data,
+                      iparms = iparms,
+                      weights = weights,
+                      index = index,
+                      parm.vector.index = parm.vector.index,
+                      replacement = replacement,
+                      root = root)
   ## optimization
   if(type == "optim"){
     op <- stats::optim(par = rep(1, length(parm.paths)), 
@@ -276,6 +289,21 @@ optim_apsimx <- function(file, src.dir = ".",
                        replacement = replacement,
                        root = root,
                        ...)    
+  }
+  
+  if(type == "ucminf"){
+    
+    op <- ucminf::ucminf(par = rep(1, length(parm.paths)), 
+                         fn = obj_fun, 
+                         parm.paths = parm.paths, 
+                         data = data, 
+                         iparms = iparms,
+                         weights = weights,
+                         index = index,
+                         parm.vector.index = parm.vector.index,
+                         replacement = replacement,
+                         root = root,
+                         ...)    
   }
   
   if(type == "nloptr"){
@@ -337,7 +365,29 @@ optim_apsimx <- function(file, src.dir = ".",
     return(op.mcmc)
   }
   
-  ans <- structure(list(rss = rss, iaux.parms = iparms, op = op, n = nrow(data),
+  post.rss <- exp(op$value) ## This is the weighted RSS
+  post.unweighted.rss <- NA    
+  ## I need the unweighted RSS for computing intervals
+  if(!is.null(list(...)$hessian)){
+    if(list(...)$hessian){
+        ## This is unweighted
+        post.lrss <- obj_fun(cfs = op$par,
+                            parm.paths = parm.paths,
+                            data = data,
+                            iparms = iparms,
+                            weights = rep(1, ncol(datami)),
+                            index = index,
+                            parm.vector.index = parm.vector.index,
+                            replacement = replacement,
+                            root = root)
+        post.unweighted.rss <- exp(post.lrss)
+    }
+  }
+  
+  ans <- structure(list(pre.rss = exp(pre.lrss), 
+                        post.rss = post.rss, ## This is weighted
+                        post.unweighted.rss = post.unweighted.rss,
+                        iaux.parms = iparms, op = op, n = nrow(data),
                         parm.vector.index = parm.vector.index),
                    class = "optim_apsim")
   return(ans)